reconfig path
